requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.25,<1.0

SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0

python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

fastapi[all]>=0.110,<1.0
structlog>=23.0,<24.0
gunicorn>=21.0,<22.0

# Added: HTTP client for Companies House + external API calls
httpx>=0.27,<1.0

# Added: Robust retries for external APIs (rate limiting/backoff)
tenacity>=8.2,<9.0
